fix: resolve 19 confirmed bug-hunt defects (no regression)
Blockers/High:
- FileOperationsManager: guard against copy/move into a file's own dir
  (self-overwrite truncation = data loss); delete partial target on cancel
- Vault: MainActivity -> FragmentActivity so BiometricPrompt attaches; a
  missing host now keeps the vault LOCKED instead of unlocking unauthenticated
- Browser: confirmation dialog gates every permanent delete (single + bulk);
  clear selection when a sheet-initiated delete/transfer is cancelled
- MediaQualityProbe: score encodes only the intrinsic metric (0L on probe
  failure, never bytes) so SmartMerge can't keep a corrupt copy over the original
- Theme: accent built with the ARGB Color(Long) overload (no .toULong crash)
- ArchiveManager: write empty-directory zip entries (no FNFE); accumulate tar
  extraction progress
- WebDAV: stop double-prepending the base path; default https / honour custom
  HTTPS ports
- FTPS: execPBSZ(0)+execPROT("P") so the data channel is encrypted
- SFTP: trust-on-first-use known_hosts; reject a CHANGED host key (MITM)
- AI: drop main-thread runBlocking DataStore reads (ANR); isEnabled async/cached

Medium/Low:
- API key stored via EncryptedSharedPreferences (plaintext migration)
- Debounced search-keystroke reload; video seek no longer snaps back;
  truncated-UTF-8 text labelled correctly

+++ app/src/main/java/com/jupiter/filemanager/ui/theme/Theme.kt
@@ -48,7 +48,7 @@ fun JupiterTheme(
         // An explicit accent override always wins over dynamic color.
         accentColorArgb != 0L -> {
             val base = if (darkTheme) JupiterDarkColors else JupiterLightColors
-            base.withAccent(Color(accentColorArgb.toULong()), darkTheme)
+            base.withAccent(Color(accentColorArgb), darkTheme)
         }
         dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
             if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
